Request router isolates code name from the request

+++ IR/transmitter/cpp/IrXmitterTests/transmit/IrSignalSenderTests.cpp
@@ -7,11 +7,11 @@ namespace
 {
 	class MockHardwareInterface : public IMockHardwareInterface
 	{
-		virtual uint64_t gpio_put_at_us(int pin_number, bool value, int time_us);
+		virtual uint64_t gpio_put_at_us(int pin_number, bool value, uint64_t time_us);
 	public:
 		size_t size() const { return Times.size(); }
 		std::vector<int> Pins;
-		std::vector<int> Times;
+		std::vector<uint64_t> Times;
 		std::vector<int> Values;
 	};
 }
@@ -66,7 +66,7 @@ static void ACodeIsTransmitted()
 	AssertDataPoint(2, 302, 1);
 }
 
-uint64_t MockHardwareInterface::gpio_put_at_us(int pin_number, bool value, int time_us)
+uint64_t MockHardwareInterface::gpio_put_at_us(int pin_number, bool value, uint64_t time_us)
 {
 	Pins.push_back(pin_number);
 	Times.push_back(time_us);
